Fix ambiguous pg_size_pretty(integer) crash in analyze_db_storage.py
Found live running WO-60's cleanup on Render: an unqualified integer
argument to pg_size_pretty is genuinely ambiguous between its bigint
and numeric overloads in Postgres (neither is the "preferred" type in
that category), so the bare `pg_size_pretty(byte_size)` in the
duplicate-images query crashed with AmbiguousFunctionError. Every other
call site in the script wraps the argument in SUM()/MAX()/a
pg_*_size() function, all of which already return bigint, so this was
the only bare case. Cast explicitly.

Also corrected STORAGE_CLEANUP_2026_08_25.md: `cd /app` doesn't exist
on the real Render shell (confirmed live, real cwd is ~/project/src),
VACUUM FULL does briefly lock the tables it processes rather than
being fully "no downtime", and the 300-500MB reclaim estimate was a
guess -- the real dry-run number measured live tonight was ~117MB
across 402 pages.

+++ scripts/analyze_db_storage.py
@@ -99,7 +99,7 @@ async def main():
             SELECT
                 etag,
                 COUNT(*) as count,
-                pg_size_pretty(byte_size) as size,
+                pg_size_pretty(byte_size::bigint) as size,
                 COUNT(*) * byte_size as total_wasted
             FROM meeting_page_thumbnails
             GROUP BY etag, byte_size
